Small Refactor to Bimtester + Diff Module (#1271)
* Refactoring of BIMTester Module

* Modification to Bimtester refactor + Diff refactor
