+++ src/ifcblenderexport/blenderbim/bim/__init__.py
@@ -17,6 +17,7 @@ if bpy is not None:
         "context": None,
         "covetool": None,
         "csv": None,
+        "diff": None,
         "bimtester": None,
         "debug": None,
         "geometry": None,
@@ -40,9 +41,6 @@ if bpy is not None:
         operator.SelectClass,
         operator.SelectType,
         operator.OpenUri,
-        operator.SelectDiffJsonFile,
-        operator.SelectDiffNewFile,
-        operator.SelectDiffOldFile,
         operator.SelectDataDir,
         operator.SelectSchemaDir,
         operator.SelectIfcFile,
@@ -106,8 +104,6 @@ if bpy is not None:
         operator.OpenView,
         operator.OpenViewCamera,
         operator.ActivateView,
-        operator.ExecuteIfcDiff,
-        operator.VisualiseDiff,
         operator.ExportClashSets,
         operator.ImportClashSets,
         operator.AddClashSet,
@@ -228,7 +224,6 @@ if bpy is not None:
         ui.BIM_PT_ifcclash,
         ui.BIM_PT_library,
         ui.BIM_PT_presentation_layers,
-        ui.BIM_PT_diff,
         ui.BIM_PT_patch,
         ui.BIM_PT_mvd,
         ui.BIM_PT_presentation_layer_data,
