+++ src/ifcblenderexport/blenderbim/bim/module/diff/ui.py
@@ -1,0 +1,40 @@
+from bpy.types import Panel
+from blenderbim.bim.ifc import IfcStore
+
+
+class BIM_PT_diff(Panel):
+    bl_label = "IFC Diff"
+    bl_idname = "BIM_PT_diff"
+    bl_options = {"DEFAULT_CLOSED"}
+    bl_space_type = "PROPERTIES"
+    bl_region_type = "WINDOW"
+    bl_context = "scene"
+
+    def draw(self, context):
+        layout = self.layout
+        layout.use_property_split = True
+
+        scene = context.scene
+        bim_properties = scene.DiffProperties
+
+        layout.label(text="IFC Diff Setup:")
+
+        row = layout.row(align=True)
+        row.prop(bim_properties, "diff_old_file")
+        row.operator("bim.select_diff_old_file", icon="FILE_FOLDER", text="")
+
+        row = layout.row(align=True)
+        row.prop(bim_properties, "diff_new_file")
+        row.operator("bim.select_diff_new_file", icon="FILE_FOLDER", text="")
+
+        row = layout.row(align=True)
+        row.prop(bim_properties, "diff_relationships")
+
+        row = layout.row()
+        row.operator("bim.execute_ifc_diff")
+
+        # TODO: show if there ifc diff operation is sucessful
+        row = layout.row(align=True)
+        row.prop(bim_properties, "diff_json_file")
+        row.operator("bim.select_diff_json_file", icon="FILE_FOLDER", text="")
+        row.operator("bim.visualise_diff", icon="HIDE_OFF", text="")
