+++ src/ifcblenderexport/blenderbim/bim/operator.py
@@ -748,98 +748,6 @@ class ExecuteIfcPatch(bpy.types.Operator):
         return {"FINISHED"}
 
 
-class SelectDiffJsonFile(bpy.types.Operator):
-    bl_idname = "bim.select_diff_json_file"
-    bl_label = "Select Diff JSON File"
-    filepath: bpy.props.StringProperty(subtype="FILE_PATH")
-
-    def execute(self, context):
-        bpy.context.scene.BIMProperties.diff_json_file = self.filepath
-        return {"FINISHED"}
-
-    def invoke(self, context, event):
-        context.window_manager.fileselect_add(self)
-        return {"RUNNING_MODAL"}
-
-
-class VisualiseDiff(bpy.types.Operator):
-    bl_idname = "bim.visualise_diff"
-    bl_label = "Visualise Diff"
-
-    def execute(self, context):
-        with open(bpy.context.scene.BIMProperties.diff_json_file, "r") as file:
-            diff = json.load(file)
-        for obj in bpy.context.visible_objects:
-            obj.color = (1.0, 1.0, 1.0, 0.2)
-            global_id = obj.BIMObjectProperties.attributes.get("GlobalId")
-            if not global_id:
-                continue
-            if global_id.string_value in diff["deleted"]:
-                obj.color = (1.0, 0.0, 0.0, 0.2)
-            elif global_id.string_value in diff["added"]:
-                obj.color = (0.0, 1.0, 0.0, 0.2)
-            elif global_id.string_value in diff["changed"]:
-                obj.color = (0.0, 0.0, 1.0, 0.2)
-        area = next(area for area in bpy.context.screen.areas if area.type == "VIEW_3D")
-        area.spaces[0].shading.color_type = "OBJECT"
-        return {"FINISHED"}
-
-
-class SelectDiffOldFile(bpy.types.Operator):
-    bl_idname = "bim.select_diff_old_file"
-    bl_label = "Select Diff Old File"
-    filepath: bpy.props.StringProperty(subtype="FILE_PATH")
-
-    def execute(self, context):
-        bpy.context.scene.BIMProperties.diff_old_file = self.filepath
-        return {"FINISHED"}
-
-    def invoke(self, context, event):
-        context.window_manager.fileselect_add(self)
-        return {"RUNNING_MODAL"}
-
-
-class SelectDiffNewFile(bpy.types.Operator):
-    bl_idname = "bim.select_diff_new_file"
-    bl_label = "Select Diff New File"
-    filepath: bpy.props.StringProperty(subtype="FILE_PATH")
-
-    def execute(self, context):
-        bpy.context.scene.BIMProperties.diff_new_file = self.filepath
-        return {"FINISHED"}
-
-    def invoke(self, context, event):
-        context.window_manager.fileselect_add(self)
-        return {"RUNNING_MODAL"}
-
-
-class ExecuteIfcDiff(bpy.types.Operator):
-    bl_idname = "bim.execute_ifc_diff"
-    bl_label = "Execute IFC Diff"
-    filename_ext = ".json"
-    filepath: bpy.props.StringProperty(subtype="FILE_PATH")
-
-    def invoke(self, context, event):
-        self.filepath = bpy.path.ensure_ext(bpy.data.filepath, ".json")
-        WindowManager = context.window_manager
-        WindowManager.fileselect_add(self)
-        return {"RUNNING_MODAL"}
-
-    def execute(self, context):
-        import ifcdiff
-
-        ifc_diff = ifcdiff.IfcDiff(
-            bpy.context.scene.BIMProperties.diff_old_file,
-            bpy.context.scene.BIMProperties.diff_new_file,
-            self.filepath,
-            bpy.context.scene.BIMProperties.diff_relationships.split(),
-        )
-        ifc_diff.diff()
-        ifc_diff.export()
-        bpy.context.scene.BIMProperties.diff_json_file = self.filepath
-        return {"FINISHED"}
-
-
 class ExportClashSets(bpy.types.Operator):
     bl_idname = "bim.export_clash_sets"
     bl_label = "Export Clash Sets"
