+++ src/ifcblenderexport/blenderbim/bim/ui.py
@@ -802,43 +802,6 @@ class BIM_PT_library(Panel):
         layout.row().prop(scene.BIMLibrary, "description")
 
 
-class BIM_PT_diff(Panel):
-    bl_label = "IFC Diff"
-    bl_idname = "BIM_PT_diff"
-    bl_options = {"DEFAULT_CLOSED"}
-    bl_space_type = "PROPERTIES"
-    bl_region_type = "WINDOW"
-    bl_context = "scene"
-
-    def draw(self, context):
-        layout = self.layout
-        layout.use_property_split = True
-
-        scene = context.scene
-        bim_properties = scene.BIMProperties
-
-        layout.label(text="IFC Diff Setup:")
-
-        row = layout.row(align=True)
-        row.prop(bim_properties, "diff_json_file")
-        row.operator("bim.select_diff_json_file", icon="FILE_FOLDER", text="")
-        row.operator("bim.visualise_diff", icon="HIDE_OFF", text="")
-
-        row = layout.row(align=True)
-        row.prop(bim_properties, "diff_old_file")
-        row.operator("bim.select_diff_old_file", icon="FILE_FOLDER", text="")
-
-        row = layout.row(align=True)
-        row.prop(bim_properties, "diff_new_file")
-        row.operator("bim.select_diff_new_file", icon="FILE_FOLDER", text="")
-
-        row = layout.row(align=True)
-        row.prop(bim_properties, "diff_relationships")
-
-        row = layout.row()
-        row.operator("bim.execute_ifc_diff")
-
-
 class BIM_PT_patch(Panel):
     bl_label = "IFC Patch"
     bl_idname = "BIM_PT_patch"
